fix(ai-infra): story 2 review — retail input gate, moved actors, refused levels
- The hidden-hint input gate is retail's Q_strnicmp(input,
  "ScriptUnhide", strlen(input)): any case-insensitive prefix passes.
- m_flNextUseTime is a saved float field; the view-only write in the
  standoff body is named as a seam.
- A refused adoption builds no entity world (no level script, no spawn)
  and fails on its own readiness arm; the navigation arm of Missing() is
  restored.
- A moved infrastructure actor moves its entity: its location becomes
  the def origin and the origin keyvalue (Source inches); the 0.5 cm
  refusal is gone.
- classname/targetname keyfield rows are display-only; ClassHintType
  drops its bool out-parameter.
- Tests: the verify declared-set pytest now runs (it was skipping), def
  comparisons check output rows, the bridge test pins the first name
  match, new Moved and Readiness cases, the gate's prefix rule.

+++ Source/ElysiumUE/Private/Substrate/ElysiumHint.h
@@ -40,7 +40,7 @@ public:
 	float TargetAngleRangeDot = 0.0f;                // +0x458 m_flTargetAngleRangeDot
 	FElysiumEntityHandle HintOwner;                  // +0x5e0 m_hHintOwner (-1 from the ctor)
 	int32 NodeId = INDEX_NONE;                       // +0x5e4 m_nNodeID; SEAM until 0018 story 3
-	double NextUseTime = 0.0;                        // +0x5ec m_flNextUseTime
+	float NextUseTime = 0.0f;                        // +0x5ec m_flNextUseTime (FIELD_TIME, a float)
 
 	// The registered classname (`ElysiumNodeEntity::HintClassname`).
 	static FName ClassName();
@@ -82,7 +82,8 @@ public:
 	void InputSetUserData(const FElysiumInputArgs& Args);
 
 	// Retail's one entry gate on `CBaseEntity::AcceptInput` (`FUN_100abc90`): a hidden entity
-	// answers true, doing nothing, to every input whose name does not begin with `ScriptUnhide`.
+	// answers true, doing nothing, to every input whose name is not a case-insensitive prefix of
+	// `ScriptUnhide` (`Q_strnicmp(input, "ScriptUnhide", strlen(input))`).
 	// SEAM for the gate itself, which this substrate does not stand globally; the hint's own inputs
 	// apply it, because a hidden hint swallowing `EnableHint` is what the shipped scripts observe.
 	bool SwallowsInput(const FElysiumInputArgs& Args) const;
